+++ src/main/java/yagu/yagu/crawler/KboGame.java
@@ -1,4 +1,4 @@
-package yagu.yagu.entity;
+package yagu.yagu.crawler;
 
 import jakarta.persistence.*;
 import lombok.Getter;
