+++ src/main/java/yagu/yagu/crawler/Status.java
@@ -1,4 +1,4 @@
-package yagu.yagu.entity;
+package yagu.yagu.crawler;
 
 public enum Status {
     SCHEDULED,
